+++ app/dashboard/(overview)/loading.tsx
@@ -1,4 +1,4 @@
-import DashboardSkeleton from "@/app/ui/skeletons";
+import DashboardSkeleton from "@/source/ui/skeletons";
 
 export default function Loading() {
     return <DashboardSkeleton />;
