+++ app/dashboard/layout.tsx
@@ -1,4 +1,4 @@
-import SideNav from '@/app/ui/dashboard/sidenav';
+import SideNav from '@/source/ui/dashboard/sidenav';
 
 export const experimental_ppr = true;
